Split bench_util.c into focused modules
