+++ bench/CMakeLists.txt
@@ -7,7 +7,15 @@ function(add_bench TGT SOURCE_FILE)
 endfunction()
 
 # bench_util: static library with pipeline benchmark infrastructure
-add_library(bench_util STATIC bench_util.c bench_util.h)
+add_library(bench_util STATIC
+    bench_util.c bench_util.h
+    sink_discard.c sink_discard.h
+    sink_throttled.c sink_throttled.h
+    sink_metering.c sink_metering.h
+    bench_zarr.c bench_zarr.h
+    bench_parse.c bench_parse.h
+    bench_report.c bench_report.h
+)
 target_include_directories(bench_util PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries(
     bench_util
