+++ bench/bench_util.c
@@ -1,615 +1,26 @@
 #include "bench_util.h"
+#include "bench_parse.h"
+#include "bench_report.h"
+#include "bench_zarr.h"
 #include "defs.limits.h"
 #include "dimension.h"
 #include "gpu/compress.h"
 #include "gpu/lod.h"
 #include "gpu/prelude.cuda.h"
-#include "ngff.h"
 #include "platform/platform.h"
-#include "store.h"
+#include "sink_discard.h"
+#include "sink_metering.h"
+#include "sink_throttled.h"
 #include "stream.cpu.h"
 #include "stream/layouts.h"
 #include "util/metric.h"
 #include "util/prelude.h"
-#include "zarr.h"
-#include "zarr/io_queue.h"
 #include "zarr/json_writer.h"
-#include "zarr/store.h"
-#include "zarr/zarr_metadata.h"
 
 #include <stdatomic.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-
-// --- Bench-internal structs (not exposed in bench_util.h) ---
-
-struct discard_shard_writer
-{
-  struct shard_writer base;
-  struct discard_shard_sink* parent;
-};
-
-struct discard_shard_sink
-{
-  struct shard_sink base;
-  struct discard_shard_writer writer;
-  size_t total_bytes;
-  size_t shards_finalized;
-};
-
-#define METER_MAX_WRITERS 32
-
-struct metering_writer
-{
-  struct shard_writer base;
-  struct shard_writer* inner;
-  struct metering_sink* parent;
-  int in_use;
-};
-
-struct metering_sink
-{
-  struct shard_sink base;
-  struct shard_sink* inner;
-  struct metering_writer writers[METER_MAX_WRITERS];
-  size_t total_bytes;
-  struct stream_metric metric;
-  struct platform_clock clock;
-};
-
-struct sink_stats
-{
-  size_t total_bytes;
-  uint64_t total_chunks; // all LOD levels, per epoch
-};
-
-// --- Throughput helpers ---
-
-double
-gb_per_s(double bytes, double ms)
-{
-  if (ms <= 0)
-    return 0;
-  return (bytes / (1024.0 * 1024.0 * 1024.0)) / (ms / 1000.0);
-}
-
-// --- Discard shard_sink for benchmarks ---
-
-static int
-discard_shard_write(struct shard_writer* self,
-                    uint64_t offset,
-                    const void* beg,
-                    const void* end)
-{
-  (void)offset;
-  struct discard_shard_writer* w = (struct discard_shard_writer*)self;
-  size_t nbytes = (size_t)((const char*)end - (const char*)beg);
-  w->parent->total_bytes += nbytes;
-  return 0;
-}
-
-static int
-discard_shard_finalize(struct shard_writer* self)
-{
-  struct discard_shard_writer* w = (struct discard_shard_writer*)self;
-  w->parent->shards_finalized++;
-  return 0;
-}
-
-static struct shard_writer*
-discard_shard_open(struct shard_sink* self, uint8_t level, uint64_t shard_index)
-{
-  (void)level;
-  (void)shard_index;
-  struct discard_shard_sink* s = (struct discard_shard_sink*)self;
-  return &s->writer.base;
-}
-
-static void
-discard_shard_sink_init(struct discard_shard_sink* s)
-{
-  *s = (struct discard_shard_sink){
-    .base = { .open = discard_shard_open },
-  };
-  s->writer = (struct discard_shard_writer){
-    .base = { .write = discard_shard_write,
-              .finalize = discard_shard_finalize },
-    .parent = s,
-  };
-}
-
-// --- Throttled shard_sink: synthetic IO pressure for measurement ---
-//
-// Same async shape as shard_pool_fs: owns its own io_queue, worker sleeps
-// proportional to bytes+latency instead of doing real IO. Exercises the
-// record_fence/wait_fence/pending_bytes path so stall timing in the GPU
-// flush pipeline can be measured on machines where the real GPU can't
-// saturate disk IO.
-
-struct throttled_shard_sink;
-
-struct throttled_shard_writer
-{
-  struct shard_writer base;
-  struct throttled_shard_sink* parent;
-};
-
-struct throttled_shard_sink
-{
-  struct shard_sink base;
-  struct throttled_shard_writer writer; // single shared writer
-  struct io_queue* queue;               // owned
-  uint64_t queued_bytes;                // main-thread only
-  _Atomic uint64_t retired_bytes;       // worker, atomic
-  _Atomic uint64_t total_bytes;         // reporting
-  uint64_t latency_ns;                  // fixed per-job cost
-  uint64_t bytes_per_sec;               // 0 = no bandwidth cap
-};
-
-struct throttled_job
-{
-  uint64_t nbytes;
-  uint64_t latency_ns;
-  uint64_t bytes_per_sec;
-  _Atomic uint64_t* retired_bytes;
-  _Atomic uint64_t* total_bytes;
-};
-
-static void
-throttled_fn(void* arg)
-{
-  struct throttled_job* j = (struct throttled_job*)arg;
-  int64_t ns = (int64_t)j->latency_ns;
-  if (j->bytes_per_sec > 0)
-    ns += (int64_t)((j->nbytes * 1000000000ull) / j->bytes_per_sec);
-  if (ns > 0)
-    platform_sleep_ns(ns);
-  atomic_fetch_add(j->retired_bytes, j->nbytes);
-  atomic_fetch_add(j->total_bytes, j->nbytes);
-}
-
-static int
-throttled_post(struct throttled_shard_sink* s, size_t nbytes)
-{
-  struct throttled_job* j = (struct throttled_job*)malloc(sizeof(*j));
-  CHECK(Error, j);
-  j->nbytes = nbytes;
-  j->latency_ns = s->latency_ns;
-  j->bytes_per_sec = s->bytes_per_sec;
-  j->retired_bytes = &s->retired_bytes;
-  j->total_bytes = &s->total_bytes;
-  if (io_queue_post(s->queue, throttled_fn, j, free)) {
-    free(j);
-    goto Error;
-  }
-  s->queued_bytes += nbytes;
-  return 0;
-
-Error:
-  return 1;
-}
-
-static int
-throttled_shard_write(struct shard_writer* self,
-                      uint64_t offset,
-                      const void* beg,
-                      const void* end)
-{
-  (void)offset;
-  struct throttled_shard_writer* w = (struct throttled_shard_writer*)self;
-  size_t nbytes = (size_t)((const char*)end - (const char*)beg);
-  return throttled_post(w->parent, nbytes);
-}
-
-static int
-throttled_shard_write_direct(struct shard_writer* self,
-                             uint64_t offset,
-                             const void* beg,
-                             const void* end)
-{
-  (void)offset;
-  struct throttled_shard_writer* w = (struct throttled_shard_writer*)self;
-  size_t nbytes = (size_t)((const char*)end - (const char*)beg);
-  return throttled_post(w->parent, nbytes);
-}
-
-static int
-throttled_shard_finalize(struct shard_writer* self)
-{
-  (void)self;
-  return 0;
-}
-
-static struct shard_writer*
-throttled_shard_open(struct shard_sink* self,
-                     uint8_t level,
-                     uint64_t shard_index)
-{
-  (void)level;
-  (void)shard_index;
-  struct throttled_shard_sink* s = (struct throttled_shard_sink*)self;
-  return &s->writer.base;
-}
-
-static struct io_event
-throttled_shard_record_fence(struct shard_sink* self, uint8_t level)
-{
-  (void)level;
-  struct throttled_shard_sink* s = (struct throttled_shard_sink*)self;
-  return io_queue_record(s->queue);
-}
-
-static void
-throttled_shard_wait_fence(struct shard_sink* self,
-                           uint8_t level,
-                           struct io_event ev)
-{
-  (void)level;
-  struct throttled_shard_sink* s = (struct throttled_shard_sink*)self;
-  io_event_wait(s->queue, ev);
-}
-
-static int
-throttled_shard_has_error(const struct shard_sink* self)
-{
-  (void)self;
-  return 0;
-}
-
-static size_t
-throttled_shard_pending_bytes(const struct shard_sink* self)
-{
-  const struct throttled_shard_sink* s =
-    (const struct throttled_shard_sink*)self;
-  uint64_t retired = atomic_load(&s->retired_bytes);
-  if (s->queued_bytes <= retired)
-    return 0;
-  return (size_t)(s->queued_bytes - retired);
-}
-
-static int
-throttled_shard_sink_init(struct throttled_shard_sink* s,
-                          uint64_t io_bw_mbps,
-                          uint64_t io_latency_us)
-{
-  *s = (struct throttled_shard_sink){ 0 };
-  s->latency_ns = io_latency_us * 1000ull;
-  s->bytes_per_sec = io_bw_mbps * 1024ull * 1024ull;
-  s->queue = io_queue_create();
-  if (!s->queue)
-    return 1;
-
-  s->base.open = throttled_shard_open;
-  s->base.record_fence = throttled_shard_record_fence;
-  s->base.wait_fence = throttled_shard_wait_fence;
-  s->base.has_error = throttled_shard_has_error;
-  s->base.pending_bytes = throttled_shard_pending_bytes;
-
-  s->writer = (struct throttled_shard_writer){
-    .base = { .write = throttled_shard_write,
-              .write_direct = throttled_shard_write_direct,
-              .finalize = throttled_shard_finalize },
-    .parent = s,
-  };
-  return 0;
-}
-
-static void
-throttled_shard_sink_teardown(struct throttled_shard_sink* s)
-{
-  if (s->queue) {
-    io_event_wait(s->queue, io_queue_record(s->queue));
-    io_queue_destroy(s->queue);
-  }
-  *s = (struct throttled_shard_sink){ 0 };
-}
-
-// --- Metering shard_sink wrapper ---
-
-static int
-metering_write(struct shard_writer* self,
-               uint64_t offset,
-               const void* beg,
-               const void* end)
-{
-  struct metering_writer* w = (struct metering_writer*)self;
-  platform_toc(&w->parent->clock);
-  int rc = w->inner->write(w->inner, offset, beg, end);
-  size_t nbytes = (size_t)((const char*)end - (const char*)beg);
-  w->parent->total_bytes += nbytes;
-  accumulate_metric_ms(&w->parent->metric,
-                       (float)(platform_toc(&w->parent->clock) * 1000.0),
-                       nbytes,
-                       0);
-  return rc;
-}
-
-static int
-metering_write_direct(struct shard_writer* self,
-                      uint64_t offset,
-                      const void* beg,
-                      const void* end)
-{
-  struct metering_writer* w = (struct metering_writer*)self;
-  platform_toc(&w->parent->clock);
-  int rc = w->inner->write_direct(w->inner, offset, beg, end);
-  size_t nbytes = (size_t)((const char*)end - (const char*)beg);
-  w->parent->total_bytes += nbytes;
-  accumulate_metric_ms(&w->parent->metric,
-                       (float)(platform_toc(&w->parent->clock) * 1000.0),
-                       nbytes,
-                       0);
-  return rc;
-}
-
-static int
-metering_finalize(struct shard_writer* self)
-{
-  struct metering_writer* w = (struct metering_writer*)self;
-  int rc = w->inner->finalize(w->inner);
-  w->in_use = 0;
-  w->inner = NULL;
-  return rc;
-}
-
-static struct shard_writer*
-metering_open(struct shard_sink* self, uint8_t level, uint64_t shard_index)
-{
-  struct metering_sink* ms = (struct metering_sink*)self;
-  struct shard_writer* inner = ms->inner->open(ms->inner, level, shard_index);
-  if (!inner)
-    return NULL;
-  for (int i = 0; i < METER_MAX_WRITERS; ++i) {
-    if (!ms->writers[i].in_use) {
-      ms->writers[i].in_use = 1;
-      ms->writers[i].inner = inner;
-      ms->writers[i].base.write_direct =
-        inner->write_direct ? metering_write_direct : NULL;
-      return &ms->writers[i].base;
-    }
-  }
-  return NULL;
-}
-
-static struct io_event
-metering_record_fence(struct shard_sink* self, uint8_t level)
-{
-  struct metering_sink* ms = (struct metering_sink*)self;
-  return ms->inner->record_fence(ms->inner, level);
-}
-
-static void
-metering_wait_fence(struct shard_sink* self, uint8_t level, struct io_event ev)
-{
-  struct metering_sink* ms = (struct metering_sink*)self;
-  ms->inner->wait_fence(ms->inner, level, ev);
-}
-
-static int
-metering_has_error(const struct shard_sink* self)
-{
-  const struct metering_sink* ms = (const struct metering_sink*)self;
-  return ms->inner->has_error(ms->inner);
-}
-
-static size_t
-metering_pending_bytes(const struct shard_sink* self)
-{
-  const struct metering_sink* ms = (const struct metering_sink*)self;
-  return ms->inner->pending_bytes(ms->inner);
-}
-
-static void
-metering_sink_init(struct metering_sink* ms, struct shard_sink* inner)
-{
-  *ms = (struct metering_sink){
-    .base = {
-      .open = metering_open,
-      .record_fence = inner->record_fence ? metering_record_fence : NULL,
-      .wait_fence = inner->wait_fence ? metering_wait_fence : NULL,
-      .has_error = inner->has_error ? metering_has_error : NULL,
-      .pending_bytes = inner->pending_bytes ? metering_pending_bytes : NULL,
-    },
-    .inner = inner,
-    .metric = { .name = "Sink", .best_ms = 1e30f },
-  };
-  for (int i = 0; i < METER_MAX_WRITERS; ++i) {
-    ms->writers[i] = (struct metering_writer){
-      .base = { .write = metering_write, .finalize = metering_finalize },
-      .parent = ms,
-    };
-  }
-}
-
-// --- Polymorphic zarr output handle ---
-//
-// Wraps store + pool + array/multiscale in a single handle for the 4-way
-// dispatch (FS/S3 x single/multiscale).
-
-struct bench_zarr_handle
-{
-  struct store* store;
-  struct zarr_array* array;   // non-NULL for single-array
-  struct ngff_multiscale* ms; // non-NULL for multiscale
-};
-
-static int
-write_intermediate_bench(const char* partial, void* ctx)
-{
-  struct store* store = (struct store*)ctx;
-  store->mkdirs(store, partial);
-  char key[4096];
-  snprintf(key, sizeof(key), "%s/zarr.json", partial);
-  return zarr_write_group(store, key, NULL);
-}
-
-static int
-bench_zarr_open_fs(struct bench_zarr_handle* z,
-                   const char* store_path,
-                   const char* array_name,
-                   const struct dimension* dims,
-                   uint8_t rank,
-                   enum dtype data_type,
-                   double fill_value,
-                   struct codec_config codec,
-                   int is_multiscale)
-{
-  *z = (struct bench_zarr_handle){ 0 };
-
-  z->store = store_fs_create(store_path, 1 /* unbuffered */);
-  CHECK(Fail, z->store);
-  z->store->mkdirs(z->store, ".");
-
-  // Write root group
-  CHECK(Fail_store, zarr_write_group(z->store, "zarr.json", NULL) == 0);
-
-  // Write intermediate groups
-  if (array_name && array_name[0]) {
-    CHECK(Fail_store,
-          zarr_for_each_intermediate(
-            array_name, write_intermediate_bench, z->store) == 0);
-    CHECK(Fail_store, z->store->mkdirs(z->store, array_name) == 0);
-  }
-
-  if (is_multiscale) {
-    struct ngff_multiscale_config mscfg = {
-      .data_type = data_type,
-      .fill_value = fill_value,
-      .rank = rank,
-      .dimensions = dims,
-      .nlod = 0,
-      .codec = codec,
-    };
-    z->ms =
-      ngff_multiscale_create(z->store, array_name ? array_name : "", &mscfg);
-    CHECK(Fail_store, z->ms);
-  } else {
-    struct zarr_array_config acfg = {
-      .data_type = data_type,
-      .fill_value = fill_value,
-      .rank = rank,
-      .dimensions = dims,
-      .codec = codec,
-    };
-    z->array = zarr_array_create(z->store, array_name ? array_name : "", &acfg);
-    CHECK(Fail_store, z->array);
-  }
-  return 0;
-
-Fail_store:
-  store_destroy(z->store);
-  z->store = NULL;
-Fail:
-  return 1;
-}
-
-static int
-bench_zarr_open_s3(struct bench_zarr_handle* z,
-                   const char* bucket,
-                   const char* prefix,
-                   const char* array_name,
-                   const char* region,
-                   const char* endpoint,
-                   double throughput_gbps,
-                   const struct dimension* dims,
-                   uint8_t rank,
-                   enum dtype data_type,
-                   double fill_value,
-                   struct codec_config codec,
-                   int is_multiscale)
-{
-  *z = (struct bench_zarr_handle){ 0 };
-
-  struct store_s3_config scfg = {
-    .bucket = bucket,
-    .prefix = prefix,
-    .region = region,
-    .endpoint = endpoint,
-    .throughput_gbps = throughput_gbps,
-  };
-  store_s3_config_set_defaults(&scfg);
-
-  z->store = store_s3_create(&scfg);
-  CHECK(Fail, z->store);
-  z->store->mkdirs(z->store, ".");
-
-  // Write root group
-  CHECK(Fail_store, zarr_write_group(z->store, "zarr.json", NULL) == 0);
-
-  // Write intermediate groups
-  if (array_name && array_name[0]) {
-    CHECK(Fail_store,
-          zarr_for_each_intermediate(
-            array_name, write_intermediate_bench, z->store) == 0);
-    CHECK(Fail_store, z->store->mkdirs(z->store, array_name) == 0);
-  }
-
-  if (is_multiscale) {
-    struct ngff_multiscale_config mscfg = {
-      .data_type = data_type,
-      .fill_value = fill_value,
-      .rank = rank,
-      .dimensions = dims,
-      .nlod = 0,
-      .codec = codec,
-    };
-    z->ms =
-      ngff_multiscale_create(z->store, array_name ? array_name : "", &mscfg);
-    CHECK(Fail_store, z->ms);
-  } else {
-    struct zarr_array_config acfg = {
-      .data_type = data_type,
-      .fill_value = fill_value,
-      .rank = rank,
-      .dimensions = dims,
-      .codec = codec,
-    };
-    z->array = zarr_array_create(z->store, array_name ? array_name : "", &acfg);
-    CHECK(Fail_store, z->array);
-  }
-  return 0;
-
-Fail_store:
-  store_destroy(z->store);
-  z->store = NULL;
-Fail:
-  return 1;
-}
-
-static struct shard_sink*
-bench_zarr_as_shard_sink(struct bench_zarr_handle* z)
-{
-  if (z->ms)
-    return ngff_multiscale_as_shard_sink(z->ms);
-  return zarr_array_as_shard_sink(z->array);
-}
-
-static int
-bench_zarr_flush(struct bench_zarr_handle* z)
-{
-  int err = 0;
-  err |= ngff_multiscale_flush(z->ms);
-  err |= zarr_array_flush(z->array);
-  return err;
-}
-
-static size_t
-bench_zarr_pending_bytes(struct bench_zarr_handle* z)
-{
-  if (z->ms)
-    return ngff_multiscale_pending_bytes(z->ms);
-  return zarr_array_pending_bytes(z->array);
-}
-
-static void
-bench_zarr_close(struct bench_zarr_handle* z)
-{
-  ngff_multiscale_destroy(z->ms);
-  zarr_array_destroy(z->array);
-  store_destroy(z->store);
-  *z = (struct bench_zarr_handle){ 0 };
-}
 
 // --- Backend dispatch helpers ---
 
@@ -665,195 +76,6 @@ bench_destroy(struct bench_handle* h)
   else
     tile_stream_gpu_destroy(h->gpu);
 }
-
-// --- Report + pipeline helpers ---
-
-static void
-print_metric_row(const struct stream_metric* m)
-{
-  if (m->count <= 0)
-    return;
-  const int N = m->count;
-  double avg_ms = (double)m->ms / N;
-  double bytes_per = m->input_bytes / N;
-  double avg_gbs = gb_per_s(m->input_bytes, (double)m->ms);
-  int has_best = m->best_ms < 1e29f;
-
-  if (has_best) {
-    double best_gbs = gb_per_s(bytes_per, (double)m->best_ms);
-    print_report("  %-12s %8.2f %8.2f %10.2f %10.2f",
-                 m->name,
-                 avg_gbs,
-                 best_gbs,
-                 avg_ms,
-                 (double)m->best_ms);
-  } else {
-    print_report(
-      "  %-12s %8.2f %8s %10.2f %10s", m->name, avg_gbs, "-", avg_ms, "-");
-  }
-}
-
-static void
-log_bench_header(const struct tile_stream_layout* layout,
-                 enum dtype dtype,
-                 struct codec_config codec,
-                 size_t max_compressed_size,
-                 size_t codec_batch_size,
-                 size_t total_bytes,
-                 size_t total_elements)
-{
-  const size_t num_epochs =
-    (total_elements + layout->epoch_elements - 1) / layout->epoch_elements;
-
-  print_report("  total:       %.2f GiB (%zu elements, %zu epochs)",
-               (double)total_bytes / (1024.0 * 1024.0 * 1024.0),
-               total_elements,
-               num_epochs);
-  print_report("  chunk:       %lu elements = %lu KiB  (stride=%lu)",
-               (unsigned long)layout->chunk_elements,
-               (unsigned long)(layout->chunk_stride * dtype_bpe(dtype) / 1024),
-               (unsigned long)layout->chunk_stride);
-  print_report("  epoch:       %lu slots, %lu MiB pool",
-               (unsigned long)layout->chunks_per_epoch,
-               (unsigned long)(layout->chunk_pool_bytes / (1024 * 1024)));
-  if (codec.id != CODEC_NONE && max_compressed_size > 0)
-    print_report("  compress:    max_output=%zu comp_pool=%zu MiB",
-                 max_compressed_size,
-                 (codec_batch_size * max_compressed_size) / (1024 * 1024));
-}
-
-static void
-print_bench_report(const struct stream_metrics* metrics,
-                   const struct tile_stream_layout* layout,
-                   enum dtype dtype,
-                   const struct sink_stats* ss,
-                   size_t total_bytes,
-                   size_t total_elements,
-                   float wall_s,
-                   float init_s,
-                   float flush_s,
-                   size_t flush_pending_bytes)
-{
-  const size_t chunk_bytes = layout->chunk_stride * dtype_bpe(dtype);
-  const size_t num_epochs =
-    (total_elements + layout->epoch_elements - 1) / layout->epoch_elements;
-  const uint64_t chunks_per_epoch =
-    ss->total_chunks ? ss->total_chunks : layout->chunks_per_epoch;
-  const size_t total_chunks = num_epochs * chunks_per_epoch;
-  const size_t total_decompressed = total_chunks * chunk_bytes;
-  const double comp_ratio =
-    total_decompressed > 0
-      ? (double)ss->total_bytes / (double)total_decompressed
-      : 0.0;
-
-  print_report("");
-  print_report("  --- Benchmark Results ---");
-  print_report("  Input:        %.2f GiB (%zu elements)",
-               (double)total_bytes / (1024.0 * 1024.0 * 1024.0),
-               total_elements);
-  print_report("  Compressed:   %.2f GiB (ratio: %.3f)",
-               (double)ss->total_bytes / (1024.0 * 1024.0 * 1024.0),
-               comp_ratio);
-  print_report("  Chunks:       %zu (%llu/epoch x %zu epochs)",
-               total_chunks,
-               (unsigned long long)chunks_per_epoch,
-               num_epochs);
-
-  print_report("");
-  print_report("  %-12s %8s %8s %10s %10s",
-               "Stage",
-               "avg GB/s",
-               "best GB/s",
-               "avg ms",
-               "best ms");
-
-  print_metric_row(&metrics->memcpy);
-  print_metric_row(&metrics->h2d);
-  print_metric_row(&metrics->scatter);
-  print_metric_row(&metrics->lod_gather);
-  print_metric_row(&metrics->lod_reduce);
-  print_metric_row(&metrics->lod_append_fold);
-  print_metric_row(&metrics->lod_morton_chunk);
-  print_metric_row(&metrics->compress);
-  print_metric_row(&metrics->aggregate);
-  print_metric_row(&metrics->d2h);
-  print_metric_row(&metrics->sink);
-
-  // Stall stats — wall-clock time the host is blocked waiting. Emitted only
-  // if any stall was observed.
-  int have_stalls =
-    metrics->flush_stall.count > 0 || metrics->kick_sync_stall.count > 0 ||
-    metrics->io_fence_stall.count > 0 || metrics->backpressure.count > 0 ||
-    metrics->max_append_ms > 0 || metrics->peak_pending_bytes > 0;
-  if (have_stalls) {
-    print_report("");
-    print_report("  --- Stall stats ---");
-    print_metric_row(&metrics->flush_stall);
-    print_metric_row(&metrics->kick_sync_stall);
-    print_metric_row(&metrics->io_fence_stall);
-    print_metric_row(&metrics->backpressure);
-    print_report("  max append ms:   %.2f", (double)metrics->max_append_ms);
-    print_report("  peak pending:    %.2f MiB",
-                 (double)metrics->peak_pending_bytes / (1024.0 * 1024.0));
-  }
-
-  double throughput_gib =
-    wall_s > 0 ? ((double)total_bytes / (1024.0 * 1024.0 * 1024.0)) / wall_s
-               : 0.0;
-  print_report("");
-  print_report("  Init time:     %.3f s", (double)init_s);
-  if (flush_pending_bytes > 0 && flush_s > 0) {
-    double flush_gib =
-      ((double)flush_pending_bytes / (1024.0 * 1024.0 * 1024.0)) /
-      (double)flush_s;
-    print_report(
-      "  Flush time:    %.3f s (%.2f GiB/s)", (double)flush_s, flush_gib);
-  } else {
-    print_report("  Flush time:    %.3f s", (double)flush_s);
-  }
-  print_report("  Wall time:     %.3f s", wall_s);
-  print_report("  Throughput:    %.2f GiB/s", throughput_gib);
-}
-
-// --- Byte-size parser: "256K", "1M", "8G" etc. ---
-
-static size_t
-parse_bytes(const char* s)
-{
-  char* end = NULL;
-  size_t val = (size_t)strtoull(s, &end, 10);
-  if (end && *end) {
-    switch (*end) {
-      case 'k':
-      case 'K':
-        val <<= 10;
-        break;
-      case 'm':
-      case 'M':
-        val <<= 20;
-        break;
-      case 'g':
-      case 'G':
-        val <<= 30;
-        break;
-    }
-  }
-  return val;
-}
-
-// (autofit adapter removed — now using
-// tile_stream_{gpu,cpu}_advise_chunk_sizes)
-
-// --- dtype helpers ---
-
-static const char* const dtype_names[] = {
-  "u8", "u16", "u32", "u64", "i8", "i16", "i32", "i64", "f16", "f32", "f64",
-};
-static const enum dtype dtype_vals[] = {
-  dtype_u8,  dtype_u16, dtype_u32, dtype_u64, dtype_i8,  dtype_i16,
-  dtype_i32, dtype_i64, dtype_f16, dtype_f32, dtype_f64,
-};
-#define NUM_DTYPES (sizeof(dtype_vals) / sizeof(dtype_vals[0]))
 
 // --- Reusable bench driver ---
 //
@@ -1318,115 +540,6 @@ Cleanup:
   if (use_throttled)
     throttled_shard_sink_teardown(&tss);
   return rc;
-}
-
-// --- CLI parsing helpers ---
-
-// Returns the index of the matching string, or n if no match.
-static int
-match_option(const char* s, const char* const* options, int n)
-{
-  for (int i = 0; i < n; ++i)
-    if (strcmp(s, options[i]) == 0)
-      return i;
-  return n;
-}
-
-static fill_fn
-parse_fill(const char* s)
-{
-  static const char* const names[] = { "xor", "zeros", "rand" };
-  static const fill_fn fns[] = { fill_xor, fill_zeros, fill_rand };
-  int i = match_option(s, names, 3);
-  if (i < 3)
-    return fns[i];
-  fprintf(stderr, "Unknown fill: %s (expected xor, zeros, rand)\n", s);
-  return NULL;
-}
-
-static int
-parse_codec(const char* s, struct codec_config* out)
-{
-  static const char* const names[] = {
-    "none", "lz4", "zstd", "blosc-lz4", "blosc-zstd"
-  };
-  static const enum compression_codec vals[] = { CODEC_NONE,
-                                                 CODEC_LZ4_NON_STANDARD,
-                                                 CODEC_ZSTD,
-                                                 CODEC_BLOSC_LZ4,
-                                                 CODEC_BLOSC_ZSTD };
-  int i = match_option(s, names, 5);
-  if (i < 5) {
-    out->id = vals[i];
-    if (out->id == CODEC_LZ4_NON_STANDARD && out->level == 0)
-      out->level = 1;
-    if (codec_is_blosc(out->id) && out->level == 0)
-      out->level = 3;
-    return 1;
-  }
-  fprintf(stderr,
-          "Unknown codec: %s (expected none, lz4, zstd, blosc-lz4, "
-          "blosc-zstd)\n",
-          s);
-  return 0;
-}
-
-static int
-parse_reduce(const char* s, enum lod_reduce_method* out)
-{
-  static const char* const names[] = { "mean",   "min",     "max",
-                                       "median", "max_sup", "min_sup" };
-  static const enum lod_reduce_method vals[] = {
-    lod_reduce_mean,
-    lod_reduce_min,
-    lod_reduce_max,
-    lod_reduce_median,
-    lod_reduce_max_suppressed,
-    lod_reduce_min_suppressed,
-  };
-  int i = match_option(s, names, 6);
-  if (i < 6) {
-    *out = vals[i];
-    return 1;
-  }
-  fprintf(stderr,
-          "Unknown reduce: %s (expected mean, min, max, median, max_sup, "
-          "min_sup)\n",
-          s);
-  return 0;
-}
-
-// --- CLI parsing: backend ---
-
-static int
-parse_backend(const char* s, enum bench_backend* out)
-{
-  static const char* const names[] = { "gpu", "cpu" };
-  static const enum bench_backend vals[] = { BENCH_GPU, BENCH_CPU };
-  int i = match_option(s, names, 2);
-  if (i < 2) {
-    *out = vals[i];
-    return 1;
-  }
-  fprintf(stderr, "Unknown backend: %s (expected gpu, cpu)\n", s);
-  return 0;
-}
-
-// --- CLI parsing: dtype ---
-
-static int
-parse_dtype(const char* s, enum dtype* out)
-{
-  int i = match_option(s, dtype_names, NUM_DTYPES);
-  if (i < (int)NUM_DTYPES) {
-    *out = dtype_vals[i];
-    return 1;
-  }
-  fprintf(stderr,
-          "Unknown dtype: %s (expected u8, u16, u32, u64, i8, i16, i32, i64, "
-          "f16, f32, f64)\n",
-          s);
-  return 0;
 }
 
 // --- CLI driver ---
