+++ bench/sink_discard.c
@@ -1,0 +1,46 @@
+#include "sink_discard.h"
+
+#include <stddef.h>
+
+static int
+discard_shard_write(struct shard_writer* self,
+                    uint64_t offset,
+                    const void* beg,
+                    const void* end)
+{
+  (void)offset;
+  struct discard_shard_writer* w = (struct discard_shard_writer*)self;
+  size_t nbytes = (size_t)((const char*)end - (const char*)beg);
+  w->parent->total_bytes += nbytes;
+  return 0;
+}
+
+static int
+discard_shard_finalize(struct shard_writer* self)
+{
+  struct discard_shard_writer* w = (struct discard_shard_writer*)self;
+  w->parent->shards_finalized++;
+  return 0;
+}
+
+static struct shard_writer*
+discard_shard_open(struct shard_sink* self, uint8_t level, uint64_t shard_index)
+{
+  (void)level;
+  (void)shard_index;
+  struct discard_shard_sink* s = (struct discard_shard_sink*)self;
+  return &s->writer.base;
+}
+
+void
+discard_shard_sink_init(struct discard_shard_sink* s)
+{
+  *s = (struct discard_shard_sink){
+    .base = { .open = discard_shard_open },
+  };
+  s->writer = (struct discard_shard_writer){
+    .base = { .write = discard_shard_write,
+              .finalize = discard_shard_finalize },
+    .parent = s,
+  };
+}
